Move `plot_winning_segment_distributions` to frameworks.dendrites for reusability.
